requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
billiard==3.3.0.23
celery==3.1.18
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-restframework==0.0.1
django-six==1.0.4
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
idna==2.10
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.17.2
kombu==3.0.37
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
screen==1.0.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
wrapt==1.12.1
xlwt==1.3.0
